Scope shared metadata cache to language server

Copilot-Session: 8d3437a1-4998-4003-9741-faf9e1fce969
